refactor: migrate LSP config to native, remove nvim-lspconfig
With deprecation warnings now kicking in for the legacy nvim-lspconfig
setup, migrated to the native config way. Since I was already
customizing most of the servers, decided to drop the plugin completely
and copy over the configs with my modifications added.

In the process, since `typescript-tools` turned out to have a dependency
on nvim-lspconfig, I dropped it for vtsls. I'd had trouble with that one
in the past, but it seems to be working well enough now. Will give it
another try.